Adding configurable # stats to keep + cleanup
Now supports passing the amount of sample data to keep in Redis.  The
default is 180 samples, which at 10 seconds/sample is half an hour's
worth of samples.

The dashboard now uses this amount when starting up to determine how
much data to request from the server.  In addition, the worker will now
clean up any excess samples from Redis.  Clean up is currently done
every time a sample is written(since Redis operations are cheap), but
this could be done less frequently if necessary.