Move runtime gem dependencies from Gemfile back to gemspec file.
This undoes partially PR https://github.com/steelThread/redmon/pull/75/commits/5cec1c412f3ee9775430cd0ef09e016a50866459

The reason is that dependencies declared in the Gemfile but not in the gemspec are not visible when including this gem in an app bundle. This means that app startup can fail because a necessary gem for redmon (e.g. haml) is not present in the bundle. The fix is to declare all necessary runtime dependencies in the gemspec, so that any bundle that includes this gem will include its dependencies as well.

Fixes https://github.com/steelThread/redmon/issues/86