Removed dependency on activesupport.
Activesupport gem is not used at all in the code as far as I can tell (it's not required in any ruby file). The dependency is unnecessary. Also having it pinned to version 3.x, which is ancient, makes this gem incompatible with Rails 4 and 5 (and Rails 3 is ancient as well).